[math] Move bit constants to :math:utils
