+++ src/modm/math/utils/bit_constants.hpp
