+++ src/modm/platform/spi/at90_tiny_mega/spi_master.cpp.in
@@ -16,7 +16,7 @@
 #include <avr/io.h>
 #include <avr/interrupt.h>
 
-#include <modm/utils/bit_constants.hpp>
+#include <modm/math/utils/bit_constants.hpp>
 #include <modm/architecture/interface/atomic_lock.hpp>
 
 // bit 7 (0x80) is used for transfer 1 byte
