+++ src/modm/utils/utils.hpp
@@ -17,4 +17,3 @@
 #include "utils/allocator.hpp"
 #include "utils/arithmetic_traits.hpp"
 #include "utils/dummy.hpp"
-#include "utils/bit_constants.hpp"
